refactor: 토큰 생성과 추출 책임을 AuthTokenProvider와 AuthTokenExtractor로 분리
