+++ src/main/java/roomescape/auth/jwt/AuthTokenExtractor.java
@@ -1,0 +1,8 @@
+package roomescape.auth.jwt;
+
+public interface AuthTokenExtractor {
+
+    String extractPayload(String token);
+
+    String extractRole(String token);
+}
