+++ src/main/java/roomescape/auth/jwt/AuthTokenProvider.java
@@ -5,8 +5,4 @@ import roomescape.member.domain.Member;
 public interface AuthTokenProvider {
 
     String createToken(Member member);
-
-    String extractPayload(String token);
-
-    String extractRole(String token);
 }
