+++ src/main/java/roomescape/auth/jwt/JwtTokenProvider.java
@@ -1,8 +1,6 @@
 package roomescape.auth.jwt;
 
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
@@ -12,7 +10,6 @@ import java.util.Base64;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.exception.custom.AuthorizationException;
 import roomescape.member.domain.Member;
 
 @Component
@@ -46,55 +43,5 @@ public class JwtTokenProvider implements AuthTokenProvider {
                 .setExpiration(validity)
                 .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
-    }
-
-    @Override
-    public String extractPayload(String token) {
-        validateValidToken(token);
-
-        return Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody()
-                .getSubject();
-    }
-
-    @Override
-    public String extractRole(String token) {
-        validateValidToken(token);
-
-        return Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody()
-                .get("role", String.class);
-    }
-
-    private void validateValidToken(final String token) {
-        validateTokenExists(token);
-        validateTokenIntegrityAndExpiration(token);
-    }
-
-    private void validateTokenExists(final String token) {
-        if (token == null || token.isBlank()) {
-            throw new AuthorizationException("로그인 토큰이 존재하지 않습니다");
-        }
-    }
-
-    private void validateTokenIntegrityAndExpiration(String token) {
-        try {
-            Jws<Claims> claims = Jwts.parserBuilder()
-                    .setSigningKey(key)
-                    .build()
-                    .parseClaimsJws(token);
-
-            if (claims.getBody().getExpiration().before(new Date())) {
-                throw new AuthorizationException("토큰이 만료 되었습니다");
-            }
-        } catch (JwtException | IllegalArgumentException e) {
-            throw new AuthorizationException("서명이 올바르지 않거나 잘못된 토큰입니다");
-        }
     }
 }
